requirements.txt
africastalking==1.2.9
asgiref==3.8.1
Authlib==1.5.2
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.2.0
coverage==7.8.0
cryptography==44.0.2
Deprecated==1.2.14
dj-database-url==2.1.0
Django==5.1.4
django-cors-headers==4.2.0
django-environ==0.10.0
django-graphql-jwt==0.3.4
django-oidc-auth==0.0.11
django-rest-auth==0.9.5
djangorestframework==3.15.2
drf-oidc-auth==3.0.0
drf-yasg==1.21.8
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
idna==3.7
inflection==0.5.1
iniconfig==2.1.0
josepy==2.0.0
jwt==1.3.1
mozilla-django-oidc==4.0.1
packaging==23.2
pillow==11.1.0
pluggy==1.5.0
promise==2.3
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
pycparser==2.22
pycrypto==2.6.1
pyjwkest==0.6.2
PyJWT==2.10.1
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.2
requests==2.32.0
schema==0.7.7
six==1.16.0
South==1.0.2
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.5.0
whitenoise==6.5.0
wrapt==1.15.0
